requirements.txt
streamlit==1.31.0
pandas==1.5.3
Pillow
openpyxl
numpy==1.24.4